// src/shared/enums/success-message.enum.ts
export enum SuccessMessage {
  SIGNUP_INITIATED = 'Signup initiated successfully',
  PHONE_VERIFIED = 'Phone number verified successfully',
  PASSCODE_SET = 'Passcode set successfully',
  PIN_SET = 'Transaction Pin set successfully',
  SIGNUP_COMPLETED = 'Signup completed successfully',
  SIGNIN_INITIATED = 'Signin initiated successfully',
  SET_NATIONALITY = 'Nationality saved successfully',
  SET_TIN = 'TIN saved successfully',
  SET_BVN = 'BVN saved successfully',
  SET_NIN = 'NIN saved successfully',
  FACE_VERIFIED = 'Face verified successfully',
  SET_BUSINESS_ADDRESS = 'Business address saved successfully',
  RESIDENTIAL_ADDRESS_SAVED = 'Residential address saved successfully',
  BUSINESS_ADDRESS_VERIFIED = 'Business address verified successfully',
  BUSINESS_DETAILS_SAVED = 'Business details saved successfully',
  BUSINESS_DOCUMENT_UPLOADED = 'Business document uploaded successfully',
  BUSINESS_VERIFY_COMPLETED = 'Business verification completed successfully',
}